requirements.txt
 python-pptx 
 streamlit
 requests
 google-generativeai
 toml